+++ CHANGELOG.md
@@ -18,6 +18,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Undo/redo/copy/cut/paste in TextArea will now work with cmd+ on supported terminals
 - In TextArea, ctrl+u will now delete a newline if the cursor is at the start
 - in TextArea alt+delete is now bound to delete word right
+- Text opacity applied to an ansi theme will now set the dim attribute if the opacity is < 50%
+
+### Fixed
+
+- Fixed text opacity with ANSI themes creating RGB colors.  
 
 ## [8.2.6] - 2026-04-13